+++ scos_actions/actions/calibrate_y_factor.py
@@ -78,8 +78,9 @@ from scipy.signal import sosfilt
 
 from scos_actions import utils
 from scos_actions.actions.interfaces.action import Action
+from scos_actions.calibration import sensor_calibration
 from scos_actions.hardware import gps as mock_gps
-from scos_actions.settings import SENSOR_CALIBRATION_FILE, sensor_calibration
+from scos_actions.settings import SENSOR_CALIBRATION_FILE
 from scos_actions.signal_processing.calibration import (
     get_linear_enr,
     get_temperature,
