+++ scos_actions/settings.py
@@ -4,33 +4,8 @@ from os import path
 from django.conf import settings
 from environs import Env
 
-from scos_actions import calibration
-
 logger = logging.getLogger(__name__)
 env = Env()
-
-
-def get_sigan_calibration(sigan_cal_file):
-    try:
-        sigan_cal = calibration.load_from_json(sigan_cal_file)
-    except Exception as err:
-        logger.error("Unable to load sigan calibration data, reverting to none")
-        logger.exception(err)
-        sigan_cal = None
-
-    return sigan_cal
-
-
-def get_sensor_calibration(sensor_cal_file):
-    """Get calibration data from sensor_cal_file and sigan_cal_file."""
-    try:
-        sensor_cal = calibration.load_from_json(sensor_cal_file)
-    except Exception as err:
-        logger.error("Unable to load sensor calibration data, reverting to none")
-        logger.exception(err)
-        sensor_cal = None
-    return sensor_cal
-
 
 logger.info("Initializing scos-actions settings")
 CONFIG_DIR = path.join(path.dirname(path.abspath(__file__)), "configs")
@@ -46,6 +21,7 @@ if not settings.configured or not hasattr(settings, "SIGAN_CALIBRATION_FILE"):
     sigan_calibration = None
 else:
     SIGAN_CALIBRATION_FILE = settings.SIGAN_CALIBRATION_FILE
+
 if not settings.configured or not hasattr(settings, "SENSOR_CALIBRATION_FILE"):
     logger.warning("Using default sensor cal file.")
     SENSOR_CALIBRATION_FILE = path.join(CONFIG_DIR, "sensor_calibration_example.json")
@@ -54,6 +30,7 @@ else:
     SENSOR_CALIBRATION_FILE = settings.SENSOR_CALIBRATION_FILE
     logger.debug("SCOS_ACTIONS: SENSOR_CALIBRATION_FILE: " + SENSOR_CALIBRATION_FILE)
 
+SWITCH_CONFIGS_DIR = env("SWITCH_CONFIGS_DIR", default=None)
 if not settings.configured:
     PRESELECTOR_CONFIG_FILE = None
     SENSOR_DEFINITION_FILE = None
@@ -62,6 +39,8 @@ if not settings.configured:
     PRESELECTOR_CLASS = env("PRESELECTOR_CLASS", default=None)
     SIGAN_POWER_CYCLE_STATES = env("SIGAN_POWER_CYCLE_STATES", default=None)
     SIGAN_POWER_SWITCH = env("SIGAN_POWER_SWITCH", default=None)
+    MOCK_SIGAN = env("MOCK_SIGAN", default=None)
+
 
 else:
     MOCK_SIGAN = settings.MOCK_SIGAN
@@ -88,11 +67,3 @@ else:
         SIGAN_POWER_SWITCH = settings.SIGAN_POWER_SWITCH
     if hasattr(settings, "SIGAN_POWER_CYCLE_STATES"):
         SIGAN_POWER_CYCLE_STATES = settings.SIGAN_POWER_CYCLE_STATES
-
-
-logger.info("Loading sensor cal file: " + SENSOR_CALIBRATION_FILE)
-sensor_calibration = get_sensor_calibration(SENSOR_CALIBRATION_FILE)
-logger.info("Loading sigan cal file: " + SIGAN_CALIBRATION_FILE)
-sigan_calibration = get_sigan_calibration(SIGAN_CALIBRATION_FILE)
-if sensor_calibration:
-    logger.info("last sensor cal: " + sensor_calibration.calibration_datetime)
